Separate `pipedream` router from `main` routes
