+++ index.js
@@ -105,10 +105,12 @@ const mainRouter = require("./routes/main.route");
 const lineRouter = require("./routes/line.route");
 const herokuRouter = require("./routes/heroku.route");
 const whatsAppRouter = require("./routes/whatsapp.route");
+const pipedreamRouter = require("./routes/pipedream.route");
 app.use("/", mainRouter);
 app.use("/line", lineRouter);
 app.use("/whatsapp", whatsAppRouter);
 app.use("/heroku", herokuRouter);
+app.use("/pipedream", pipedreamRouter);
 
 app.listen(PORT, () => {
   console.log(`Server is running on PORT ${PORT}`);
